Module type S.DB

type t

module KeyMap : Stdlib.Map.S with type KeyMap.key = string

val read : t -> string -> string option

read t key is the last known value of key.

val write : t -> string -> string -> unit Lwt.t

write t key value sets key to value.

val bindings : t -> string KeyMap.t

bindings t is the bindings of t at the time of the call.

val after : t -> string KeyMap.t -> string KeyMap.t Lwt.t

after prev waits until the current bindings are different to prev and then returns the new bindings.